Move save_ooi into SingleOOIMixin
